Add Identity

+++ FitnessWebApp/Controllers/CaloriesController.cs
@@ -1,16 +1,21 @@
 using FitnessWebApp.Data;
 using FitnessWebApp.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
 namespace FitnessWebApp.Controllers;
 
+[Authorize]
 public class CaloriesController : Controller
 {
     private readonly AppDbContext _context;
+    private readonly UserManager<IdentityUser> _userManager;
 
-    public CaloriesController(AppDbContext context)
+    public CaloriesController(AppDbContext context,  UserManager<IdentityUser> userManager)
     {
         _context = context;
+        _userManager = userManager;
     }
 
     public IActionResult Index([FromQuery()] string? date = null)
@@ -20,7 +25,14 @@ public class CaloriesController : Controller
         {
             selectedDate = dateParsed;
         }
-        var caloriesLog = _context.CaloriesLogs.Where(c => c.Date == selectedDate).ToList();
+        
+        var userId = _userManager.GetUserId(User);
+        if (userId == null)
+        {
+            return Unauthorized();
+        }
+        
+        var caloriesLog = _context.CaloriesLogs.Where(c => c.Date == selectedDate).Where(c => c.UserId == userId).ToList();
         var totalCalories = caloriesLog.Sum(c => c.Calories);
         return View(new CaloriesIndexViewModel { CaloriesLogs = caloriesLog, SelectedDate = selectedDate,  TotalCalories = totalCalories });
     }
@@ -48,7 +60,14 @@ public class CaloriesController : Controller
         {
             return View(caloriesLog);
         }
-
+        
+        var userId = _userManager.GetUserId(User);
+        if (userId == null)
+        {
+            return Unauthorized();
+        }
+        
+        caloriesLog.UserId = userId;
         _context.CaloriesLogs.Add(caloriesLog);
         _context.SaveChanges();
         
@@ -59,6 +78,18 @@ public class CaloriesController : Controller
     public IActionResult Edit(int id)
     {
         var caloriesLog = _context.CaloriesLogs.Find(id);
+        
+        var userId = _userManager.GetUserId(User);
+        if (caloriesLog == null)
+        {
+            return NotFound();
+        }
+        
+        if (userId == null || caloriesLog.UserId != userId)
+        {
+            return Unauthorized();
+        }
+        
         return View(caloriesLog);
     }
     
@@ -69,7 +100,28 @@ public class CaloriesController : Controller
         {
             return View(caloriesLog);
         }
-        _context.CaloriesLogs.Update(caloriesLog);
+        
+        if (id != caloriesLog.Id)
+        {
+            return BadRequest();
+        }
+        
+        var userId = _userManager.GetUserId(User);
+        var caloriesLogDb = _context.CaloriesLogs.Find(id);
+        
+        if (caloriesLogDb == null)
+        {
+            return NotFound();
+        }
+        
+        if (userId == null ||  caloriesLogDb.UserId != userId)
+        {
+            return Unauthorized();
+        }
+        
+        caloriesLogDb.Name = caloriesLog.Name;
+        caloriesLogDb.Calories = caloriesLog.Calories;
+        
         _context.SaveChanges();
         return RedirectToAction("Index", "Calories", new { date = caloriesLog.Date.ToString("yyyy-MM-dd") });
     }
@@ -77,10 +129,16 @@ public class CaloriesController : Controller
     [HttpPost("delete/{id}")]
     public IActionResult Delete(int id)
     {
+        var userId = _userManager.GetUserId(User);
         var caloriesLog = _context.CaloriesLogs.Find(id);
         if (caloriesLog == null)
         {
             return NotFound();
+        }
+
+        if (caloriesLog.UserId != userId)
+        {
+            return Unauthorized();
         }
         _context.CaloriesLogs.Remove(caloriesLog);
         _context.SaveChanges();
